Extract CollectMachineMetric and put under behaviour
